roboweave_cloud_agent/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "roboweave-cloud-agent"
version = "0.1.0"
description = "Cloud-side LLM/VLM Agent for RoboWeave task decomposition and failure analysis"
requires-python = ">=3.10"
dependencies = [
    "roboweave-interfaces",
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "protobuf>=4.25.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "hypothesis>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["roboweave_cloud_agent*"]
